Update FromConveyorSchemaProviderTest.php
Perhaps
https://stackoverflow.com/questions/28865066/phpunit-fails-asserting-that-two-arrays-are-equal-but-shows-no-difference

has relevance here.

"Found the issue. Elements were not in the same order. A bit annoying though that the output doesn't show it."

"I have replaced assertEquals with:

$this->assertTrue($this->similar_arrays($arr1, $arr2));

protected function similar_arrays($a, $b) {
    if(is_array($a) && is_array($b)) {
        if(count(array_diff(array_keys($a), array_keys($b))) > 0)
            return false;

        foreach($a as $k => $v) {
            if(!$this->similar_arrays($v, $b[$k]))
                return false;
        }

        return true;
    }
    else
        return $a === $b;
}"

The test passes when 'title' => 'string' is inserted so that the arrays are equal in terms of 'key-value pair' count and not in the same order.